+++ CHANGELOG.md
@@ -6,8 +6,10 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+- Add ``useInstanceGranularity`` option for marker, transparency, clipping, overpaint, substance data to save memory
 - CellPack extension tweaks
     - Use instancing to create DNA/RNA curves to save memory
+    - Enable ``useInstanceGranularity`` by default
 
 ## [v3.9.0] - 2022-05-30